requirements.txt
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.7
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
